Add duplicate resolution on query handler registration, defaulting to logging. Rename packages
