+++ messaging/src/main/java/org/axonframework/queryhandling/SimpleQueryBus.java
@@ -30,8 +30,8 @@ import org.axonframework.messaging.unitofwork.DefaultUnitOfWork;
 import org.axonframework.messaging.unitofwork.UnitOfWork;
 import org.axonframework.monitoring.MessageMonitor;
 import org.axonframework.monitoring.NoOpMessageMonitor;
-import org.axonframework.queryhandling.duplication.DuplicateQueryHandlerResolution;
-import org.axonframework.queryhandling.duplication.DuplicateQueryHandlerResolver;
+import org.axonframework.queryhandling.registration.DuplicateQueryHandlerResolution;
+import org.axonframework.queryhandling.registration.DuplicateQueryHandlerResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Mono;
